Start of Breaking fix to make this branch more compliant with ModBus spec
Refer to BranchNotes.txt

Paused this branch as still no response for other PR's in upstream Repo.
Might be best before continuing to raise issue with Repo Devs if they think this would be a PR they would merge as unless leave duplicates of previous terminology & methods (messy) it would be a breaking fix.
Using this branch locally as it fixes enough for our retirements so far. ie high speed polling.